fix(scheduler,napcat): SSE 背压保护抽到 @kagami/http/sse 共用
createBackpressureAwareWrite 从 napcat 抽到 @kagami/http/sse，scheduler tick 端点复用：
慢/半死消费方背压即停写、宽限期后 destroy 让对端重连，防进程内存无界增长（补 #425 遗漏的 scheduler）。
